+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `data-ignored` boolean attribute for detecting elements, that must not be checked for structure difference.
 - `state` getter and `changeState` method to `WebComponent` class.
 - Order of render events (multiple invokation of `Router.to`, `Router.reload`, multiple state changing accomplish render in order they are called).
+- `Function` type to `State` interface.
 
 ### Changed